fix: FormBuilderCompanionMixin.canSubmit work
Because FormBuilder.of(context) does not add dependency to
the Form (_FormScope) instead of Form.of(context). It causes missing
rebuild after validation, thus canSubmit is not called after validation.
In fact, FormBuilder uses Form to implement itself maybe because of
FormBuilder also uses FormField to work their FomrBuilderFields

So, by just calling Form.of(context), we can add depencency to
_FormScope from current BuildContext, and then rebuild after validation
is enabled and canSubmit should work correctly.