BREAKING_CHANGE: always validate fields on submit
The form can be submitted initially even when there are any invalid
values in default values because no validations are occurred if
AutovalidateMode.onUserInteraction is specified.
So, we should call validateAndSave() in submit(...) method to ensure
that all properties are valid in doSubmit() method.

This commit also removes BuildContext parameter from doSubmit()
because:
* No usecalses are found.
* BuildContext is invalidated on calling doSubmit() because submit()
  now calls async validateAndSave() method.