fix: keep continuous call order
The previous implementation naively check that whether _processingValue
becomes null or not to determine whether the awaited operation has been
canceled or not. This assumption is broken when multiple invocation is
occurred during previous invocations have not been canceled (completed)
yet. Because the last invocation set _processingValue to non-null value,
and then some invocation, which might be canceled, is completed, it
looks_processingValue is not null, so the result is wrongly adapted as
the result of last, not-canceled invocation.

This commit fixes the condition to check equality between
_processingValue and parameter to ensure that the completed invocation
is certainly last, not-canceled invocation.