chore: separate tests
Flutter test tools shows bizarre error when and only when runs the
widgets test at once. The error is says that closure is not sub-type of
function, even if there are no compilation errors, and no runtime errors
when we run only subset of them.
We think it is caused by runtime when there are too many widget test
cases in the same isolate. So, we decide to separate tests to two files
which have own main() and will be executed in different isolates.