fix: async validator error causes hang-up
This commit fixes error handling for synchronous exception causes
hang-up because it calls chain terminator with isSync: false, so it
will decrement the counter and causes assertion error. In this case,
completion handler should not decrement the counter.

We fix this bug with adding a flag to indicate whether async operations
has been started, and passing the reversed flag to `isSync` parameter of
terminator call.