test: remove workaround
As of PR 1019 of `flutter_form_builder` solves the problem, so this
commit removes workarounds from tests.

Note that there is a incompatibility between `TextFormField` and
`FormBuilderTextField` when they are validated for `null` initial value.
I believe that `FormBuilderTextField`'s behavior is more appropriate
but it is hard to change `TextFormField`'s behavior, so it is reasonable
to handle the difference in the test logic with notable comments.