chore: fix restoration failure
This fix is 'chore' because it is not released yet.

Restoration failure should not throw because if compatibility failure
is occurred, it is good to recover from it. So, fallback to default value
is reasonable for form field inputs.
Note that, restoration from old version app should not be occurred.